Upload the generated word list and affix candidates as an artifact
The https://github.com/brown-uk/dict_uk project does not include a
simple word list in text format in its releases, but explains how
to build it in README.md (see the "How to run" section).

This github CI action builds the word list and uploads it as an
artifact. Additionally, the 'findaffix' tool from Ispell is used
to generate possible word prefix/suffix candidates, which can be
used for automatically constructing a semi-decent .AFF file
for Hunspell.